Fixing Live chat
